% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{antiparallel}
\alias{antiparallel}
\title{Find the antiparallel sequence}
\usage{
antiparallel(x)
}
\arguments{
\item{x}{A character sequence of bases (A, T, C, G)}
}
\value{
A character sequence with the antiparalell sequence of \code{x}
}
\description{
Find the antiparallel sequence
}
\examples{
seq <- "ATCG"
antiparallel(seq)

seq2 <- c(seq, "AUGDF")
antiparallel(seq2)
}
