% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_rna.R
\name{dna_rna}
\alias{dna_rna}
\alias{dna_to_rna}
\alias{rna_to_dna}
\title{Tools for transcription DNA to RNA}
\usage{
dna_to_rna(dna, type = "code")

rna_to_dna(rna)
}
\arguments{
\item{dna, rna}{A character vector containing the code sequence (e.g., "ACG")}

\item{type}{The type of sequence. It must be either \code{"code"} (the code helix)
or \code{"mold"} the helix read by RNA polymerase.}
}
\value{
\itemize{
\item \code{dna_to_rna()} returns a list with the following elements:
\itemize{
\item \code{c_dna} The 'code' DNA.
\item \code{m_dna} The molder DNA (antiparallel of \code{c_dna})
\item \code{m_rna} The pré- messenger RNA, i.e., the \code{c_dna} with \code{U} instead \code{T}.
}
\item \code{rna_to_dna()} returns a list with the following elements:
\itemize{
\item \code{m_rna} The messenger RNA declared.
\item \code{m_dna} The molder DNA (antiparallel of \code{m_rna})
\item \code{c_rna} The 'code' DNA., i.e., the \code{m_rna} with \code{T} instead \code{U}.
}
}
}
\description{
Tools for transcription DNA to RNA
}
\examples{
# convert a dna (code or mold) to a messenger RNA
dna_to_rna(c("ATGCGC", "AGC"))
dna_to_rna(c("ATGCGC", "AGC"), type = "mold")

# convert a messenger RNA to dna code and mold
rna_to_dna(c("AUGCGC", "CGACGGA"))
}
