% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{highlight_pattern}
\alias{highlight_pattern}
\title{Highlight a pattern sequence in a vector}
\usage{
highlight_pattern(x, pattern, replacement = "_", verbose = TRUE)
}
\arguments{
\item{x}{A character vector.}

\item{pattern}{A character vector indicating the pattern to highlight}

\item{replacement}{A replacement to replace all the characters that do not
match the \code{pattern} with.}
}
\value{
A list with the location and highlighted pattern.
}
\description{
The function highlights a pattern (e.g., "UAG") in a vector by putting a
replacement  (e.g., an underscore _) in all characters that don't match the
sequence
}
\examples{
x <- rand_nuc_seq(n = 50, seed = 1)
x
highlight_pattern(x, pattern = "CCA")

# or a vector
highlight_pattern(x, pattern = c("CCA", "GTA"))

# vectorized

x2 <-  rand_nuc_seq(n = 50, seed = 2)
x2
highlight_pattern(c(x, x2), pattern = "ATG")
}
