\name{import.gene.summary}
\alias{import.gene.summary}
\title{import gene.summary file}
\usage{
  import.gene.summary(file = "/biomirror/mouse/Gene/gene.summary.mm.txt")
}
\arguments{
  \item{file}{the path to a gene.summary file.}
}
\value{
  \code{data.frame}
}
\description{
  all the other NCBI files have an import function, so why
  not this one?
}
\author{
  Mark Cowley, 19/3/07
}

