\name{import.gene2accession}
\alias{import.gene2accession}
\title{import.gene2accession}
\usage{
  import.gene2accession(file = "/biomirror/ftp.ncbi.nih.gov/gene/DATA/gene2accession.gz",
    taxid = 10090, make.unique = F)
}
\arguments{
  \item{file}{path to gene2accession.gz}

  \item{taxid}{numeric taxon id}

  \item{make.unique}{logical}
}
\value{
  undocumented
}
\description{
  import.gene2accession
}
\author{
  Mark Cowley, 2013-05-30
}

