\name{import.gene2refseq}
\alias{import.gene2refseq}
\title{import a local copy of the gene2refseq file.}
\usage{
  import.gene2refseq(file = "/biomirror/ftp.ncbi.nih.gov/gene/DATA/gene2refseq.gz",
    taxid = 10090, make.unique = T)
}
\description{
  import a local copy of the gene2refseq file.
}
\author{
  Mark Cowley, 10 April 2006
}
\seealso{
  \code{\link{import.ncbi.file}}
}

