\name{import.gene_info}
\alias{import.gene_info}
\title{import a local copy of the gene_info file.}
\usage{
  import.gene_info(file = "/biomirror/ftp.ncbi.nih.gov/gene/DATA/gene_info.gz",
    taxid = 10090, add.cM.column = T)
}
\description{
  import a local copy of the gene_info file.
}
\author{
  Mark Cowley, 10 April 2006
}
\seealso{
  \code{\link{import.ncbi.file}}
}

