\name{import.miRbase.miRome}
\alias{import.miRbase.miRome}
\title{import.miRbase.miRome}
\usage{
  import.miRbase.miRome(f = "~/data/miRBase/sequences/latest/miRNA.csv",
    species = "hsa")
}
\arguments{
  \item{f}{the lastest miRBase miRNA.csv file (from Sanger
  FTP)}

  \item{species}{which 3 letter species code? (default
  hsa)}
}
\description{
  Obtain the set of all mature 'miRs' from miRBase = the
  'miR-ome'
}
\author{
  Mark Cowley, 2008-09-22
}

