\name{import.unigene.data}
\alias{import.unigene.data}
\title{Import the Mm.data.gz file file.}
\usage{
  import.unigene.data(taxid = TAXID.MOUSE,
    files = list(mouse = "/biomirror/bio-mirror.grangenet.net/biomirror/unigene/Mus_musculus/Mm.data.gz", human = "/biomirror/bio-mirror.grangenet.net/biomirror/unigene/Homo_sapiens/Hs.data.gz", rat = "/biomirror/bio-mirror.grangenet.net/biomirror/unigene/Rattus_norvegicus/Rn.data.gz"),
    genbank.ids = NULL)
}
\arguments{
  \item{taxid}{numeric taxon id}

  \item{files}{list of named paths to Hs.data.gz,
  Mm.data.gz, Rn.data.gz}

  \item{genbank.ids}{undocumented

  return undocumented}
}
\description{
  Import the Mm.data.gz file file.
}
\author{
  Mark Cowley, 30/3/07
}

