\name{is.ucsc.string}
\alias{is.ucsc.string}
\title{is a UCSC string?}
\usage{
  is.ucsc.string(x)
}
\arguments{
  \item{x}{a vector of UCSC strings}
}
\value{
  logical vector if the elements look like valid UCSC
  locations
}
\description{
  is a UCSC string?
}
\author{
  Mark Cowley
}
\seealso{
  \code{\link{make.ucsc.string}}
}

