\name{linkout.gene}
\alias{linkout.gene}
\title{Generate a weblink for a given Entrez GeneID to the Entrez gene NCBI
database.}
\usage{
  linkout.gene(x, g2r)
}
\arguments{
  \item{x}{vector of Entrez Gene ID's}

  \item{g2r}{ignored.}
}
\value{
  character vector of URL's
}
\description{
  Generate a weblink for a given Entrez GeneID to the
  Entrez gene NCBI database.
}
\author{
  Mark Cowley, 10 April 2006
}

