\name{linkout.ucsc}
\alias{linkout.ucsc}
\title{generate a weblink for a given Entrez GeneID to the UCSC genome browser.}
\usage{
  linkout.ucsc(x, g2r)
}
\arguments{
  \item{x}{vector of Entrez Gene ID's}

  \item{g2r}{undocumented}
}
\value{
  character vector of URL's
}
\description{
  generate a weblink for a given Entrez GeneID to the UCSC
  genome browser.
}
\author{
  Mark Cowley, 10 April 2006
}

