% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{nth_element}
\alias{nth_element}
\title{Extract the nth element of a numeric or character sequence}
\usage{
nth_element(x, interval)
}
\arguments{
\item{x}{A numeric or character sequence}

\item{interval}{The nth  interval to select from the sequence}
}
\value{
A character or numeric vector depending on the \code{x} input
}
\description{
Extract the nth element of a numeric or character sequence
}
\examples{
num <- 1:20
nth_element(num, interval = 5)

let <- letters[1:20]
nth_element(let, interval = 2)
}
