% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{rand_acg}
\alias{rand_acg}
\title{Generate random codons with A, C, and G only}
\usage{
rand_acg(n)
}
\arguments{
\item{n}{The number of random codons to generate.}
}
\value{
A character vector with the same length of \code{n}
}
\description{
Generate random codons with A, C, and G only
}
\examples{
rand_acg(n = 5)
}
