% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rand_nuc_seq.R
\name{rand_nuc_seq}
\alias{rand_nuc_seq}
\title{Generate random nucleotide sequences}
\usage{
rand_nuc_seq(n = 30, type = "dna", seed = NULL)
}
\arguments{
\item{n}{The number of nucleotides.}

\item{type}{The type of sequence. It must be either \code{"dna"} to generate a DNA
sequence, or \code{"rna"} to generate a RNA sequence}

\item{seed}{The seed for the random number generator . Set to ensure
reproducibility}
}
\value{
A character vector.
}
\description{
Generate random sequences of nucleotides (i.e., A, T, C, and G) for DNA and A,
U, C, and G for RNA. (See details for more information).
}
\details{
The random sequences are generated so that a start codon (AUG for RNA or ATG
for DNA) is inserted between the first and third bases in the chain. A stop
codon (UAA, UAG, or UGA) is randomly inserted at the end of the chain. A
message is returned with the position and stop codon used.
}
\examples{
# generate a DNA sequenc
rand_nuc_seq(n = 15, seed = 10)

# generate a RNA sequenc
rand_nuc_seq(n = 15, type = "rna", seed = 10)

}
