\name{rank.refseq}
\alias{rank.refseq}
\title{rank RefSeq identifiers}
\usage{
  rank.refseq(x)
}
\arguments{
  \item{x}{undocumented}
}
\value{
  undocumented
}
\description{
  From the tmp Handbook:
  http://www.tmp.nlm.nih.gov/books/bv.fcgi?rid=handbook.table.697
}
\details{
  I've placed INFERRED as stronger evidence than PREDICTED,
  GA or WGS since there is evidence in other organisms that
  is exists. VALIDATED >> REVIEWED >> PROVISIONAL >>
  INFERRED >> PREDICTED = GENOME ANNOTATION = WGS
}
\author{
  Mark Cowley
}

