\name{rename.gnf.markers}
\alias{rename.gnf.markers}
\title{rename.gnf.markers}
\usage{
  rename.gnf.markers(x)
}
\description{
  Simple function to rename Gnf markers to their proper
  name: eg: rename 02.118.650 -> S02Gnf118.650 It's kinda
  slow, so avoid calling it in a loop
}
\author{
  Mark Cowley, 22 March 2006
}

