% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/screenBlast.R
\name{screen_Blast}
\alias{screen_Blast}
\title{Screen within a reference sequence the presence of a querry sequence}
\usage{
screen_Blast(subject, querry, pc_id_treshold, path_blastn)
}
\arguments{
\item{subject}{The path of the reference sequence that you want to screen. The sequence should be a .fasta file including one or several sequences}

\item{querry}{The path of the querry sequence. he sequence should be a .fasta file including one or several sequences}

\item{pc_id_treshold}{the minimum percentage of identity obtained with Blast that should be obtained in order to consider that the query match the reference sequence}

\item{path}{the directory where Blastn is located}
}
\value{
numeric value of the percentage of the reference sequence which is covered by the querry sequence.
}
\description{
With this function, the user provides a reference sequence and a querry sequence and the function compute the percentage of the reference which is covered by the querry. This computation is computed using the GenomicRanges Bioconductor objects.
}
