% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/screenBlast_mlst.R
\name{screen_Blast_mlst}
\alias{screen_Blast_mlst}
\title{Screen within a reference sequence the presence of a querry sequence}
\usage{
screen_Blast_mlst(subject, querry, path_blastn, profile)
}
\arguments{
\item{subject}{The path of the reference sequence that you want to screen. The sequence should be a .fasta file including one or several sequences}

\item{querry}{The path of the querry sequence. he sequence should be a .fasta file including one or several sequences}

\item{path_blastn}{}
}
\value{
numeric value of the percentage of the reference sequence which is covered by the querry sequence.
}
\description{
With this function, the user provides a reference sequence and a querry sequence and the function compute the percentage of the reference which is covered by the querry. This computation is computed using the GenomicRanges Bioconductor objects.
}
