% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{split_each}
\alias{split_each}
\title{Split a character sequence systematically each \code{n} pieces}
\usage{
split_each(x, n = 3)
}
\arguments{
\item{x}{A character sequence/vector}

\item{n}{The number of pieces for each split}
}
\value{
A character vector
}
\description{
Split a character sequence systematically each \code{n} pieces
}
\examples{
split_each("TESTSPLITEACH2", n = 2)

# split a character vector each 3 letter
split_each(c("AAAAAABBBBBBB", "BAFASFDEASDFS"), n = 3)

}
