% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{tidy_sequence}
\alias{tidy_sequence}
\title{Tidy up a character vector}
\usage{
tidy_sequence(x)
}
\arguments{
\item{x}{A character vector.}
}
\value{
A character vector of the same length as \code{x}
}
\description{
The function removes all spaces, punctuation signals, and numbers from a
character vector, returning a 'letter only' upper-case sequence
}
\examples{
messy_seq <- c("a 1, -C B", "AS.DF2D F33")
tidy_sequence(messy_seq)
}
