% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/translate.R
\name{translate}
\alias{translate}
\title{Translate a RNA sequence to a protein sequence}
\usage{
translate(
  mrna,
  transl_table = 1,
  start_codon = "AUG",
  start_codon_first = TRUE,
  no_stop = FALSE,
  type = "olc",
  sep = ""
)
}
\arguments{
\item{mrna}{The messenger RNA.}

\item{transl_table}{The translation table used. Defaults to 1 (standard
genetic code). See more at \code{\link[=aminoacid]{aminoacid()}}}

\item{start_codon}{The start codon. Defaults to \code{"AUG"}.}

\item{start_codon_first}{Start the translation with the \code{start_codon}?
defaults to \code{TRUE}. In this case, all bases before the \code{start_codon} will
be ignored.}

\item{no_stop}{A logical indicating if the translation shouldn't stop when a
stop codon is encountered. Defaults to FALSE. If \code{TRUE}, translation will
continue up to the end of the chain, and stop codons will be represented by
\code{"*"}.}

\item{type}{The type of output. One of
\itemize{
\item \code{"olc"} for one-letter code, e.g., "V" (default).
\item \code{"tlc"} for three-letter code, e.g., "Val".
\item \code{"name"} for the complete name, e.g., "Valine".
}}

\item{sep}{The separator between amino acids. Defaults to \code{"-"}.}
}
\value{
A character
}
\description{
Translate a nucleotide (RNA) sequence to a protein chain according to the
standard genetic code
}
\examples{
code <- rand_nuc_seq(25, type = "rna")
translate(code)
}
