\name{ucsc.pos.add}
\alias{ucsc.pos.add}
\title{ucsc.pos.add}
\usage{
  ucsc.pos.add(x, bases)
}
\description{
  Shift a set of UCSC-style intervals a certain number of
  bases along the same chromosome. This can be + or -, but
  no error checking is done, only warnings made if start
  locations become negative.
}
\author{
  Mark Cowley, 2009-01-08
}

