\name{ucsc.pos.split}
\alias{ucsc.pos.split}
\title{ucsc.pos.split}
\usage{
  ucsc.pos.split(x)
}
\arguments{
  \item{x}{character vector of UCSC-style locations}
}
\value{
  a 3 column \code{data.frame} 'chr', 'start', 'stop'
}
\description{
  split UCSC-style locations to \code{data.frame} style
}
\author{
  Mark Cowley, 2013-05-30
}

