\name{ucsc.size}
\alias{ucsc.size}
\title{interval sizes}
\usage{
  ucsc.size(x)
}
\arguments{
  \item{x}{a vector of ucsc-style location strings}
}
\value{
  an integer vector of interval sizes
}
\description{
  Calculate the size of each genomic interval listed in a
  ucsc-style string
}
\author{
  Mark Cowley, 2011-08-02
}

