\name{ucsc.union}
\alias{ucsc.union}
\title{union of 2 genomic intervals}
\usage{
  ucsc.union(x, y)
}
\arguments{
  \item{x}{a vector of ucsc-style location strings}

  \item{y}{a vector of ucsc-style location strings}
}
\value{
  the union of the 2 intervals
}
\description{
  union of 2 genomic intervals
}
\author{
  Mark Cowley, 2011-08-02
}

