\name{ucsc.unionN}
\alias{ucsc.unionN}
\title{union of multiple genomic intervals}
\usage{
  ucsc.unionN(...)
}
\arguments{
  \item{\dots}{at least 2 vectors of ucsc-style location
  strings}
}
\value{
  the union of the genomic intervals
}
\description{
  union of multiple genomic intervals
}
\author{
  Mark Cowley, 2011-08-02
}

