% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{util_is}
\alias{util_is}
\alias{is_dna}
\alias{is_rna}
\title{Check if a sequence is DNA or RNA}
\usage{
is_dna(x, verbose = TRUE)

is_rna(x, verbose = TRUE)
}
\arguments{
\item{x}{A character vector}

\item{verbose}{Logical value indicating if a warning message should be
returned.}
}
\value{
A logical value indicating if a warning should be returned
}
\description{
\code{is_dna()} checks the vector \code{x} and returns \code{FALSE} if a Uracil (U) is found
in the sequence.
\code{is_rna()} checks the vector \code{x} and returns \code{FALSE} if a Thymine (T) is
found in the sequence.
}
\examples{
is_dna("ATG")
is_rna(c("AUG", "TGA"))
}
